code re-org
